requirements.txt
Django==4.1.7
djangorestframework==3.14.0
pymongo==3.12.3
sqlparse==0.2.4
dnspython==2.3.0
jsonschema==4.17.3
arrow==1.2.3 
fqdn==1.5.1 
idna==3.4 
isoduration==20.11.0 
jsonpointer==2.3 
python-dateutil==2.8.2 
rfc3339-validator==0.1.4 
rfc3987==1.3.8 
six==1.16.0 
uri-template==1.2.0 
webcolors==1.13
python_dotenv==1.0.0
django-cors-headers==3.14.0
python-magic==0.4.27
python-magic-bin==0.4.14